Update index.html
News 4 U is a website targeted to teenagers. It is meant to educate teenagers on current issues and also ways for them to get involved in their own cities. The website includes a variety of clickable images that feature articles. It also has a map and a contact form to make the website more interactive.